Moved DictionaryArray to its own module.
